refactor(functions): split functions.js Split functions.js into separate files so that it is much easier to orient yourself in
