+++ apps/clientside-bot/src/utils/functions/sendGuildMessage.ts
@@ -1,0 +1,36 @@
+import { Guild, TextChannel } from "discord.js"
+
+export type ArgumentTypes<F> = F extends (...args: infer A) => any ? A : never
+
+export default async function sendGuildMessage(
+	guild: Guild,
+	message: ArgumentTypes<TextChannel["send"]>[0]
+) {
+	const owner = () => {
+		guild
+			.fetchOwner()
+			.then((owner) => owner.send(message))
+			.catch(() => {
+				console.log(
+					`Could not send message to guild ${guild.name} (${guild.id})`
+				)
+			})
+	}
+
+	const systemChannel = () => {
+		// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+		guild.systemChannel!.send(message).catch(() => owner())
+	}
+
+	const publicUpdatesChannel = () => {
+		// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+		guild.publicUpdatesChannel!.send(message).catch(() => systemChannel())
+	}
+	if (guild.publicUpdatesChannel) {
+		publicUpdatesChannel()
+	} else if (guild.systemChannel) {
+		systemChannel()
+	} else {
+		owner()
+	}
+}
