+++ apps/clientside-bot/src/utils/functions/splitIntoGroups.ts
@@ -1,0 +1,22 @@
+/**
+ * Split a single large array into multiple smaller ones
+ * @param items A single array of items to split into smalleer groups of the specified size
+ * @param maxCount Maximum items in a group
+ * @returns An array of arrays of items, with the maximum number of items per group being the specified size
+ */
+export default function splitIntoGroups<T>(items: T[], maxCount = 500): T[][] {
+	return items.reduce<T[][]>(
+		(previous, item) => {
+			const last = previous[previous.length - 1]
+			if (last.length >= maxCount) {
+				// if the last group is full, start a new one
+				previous.push([item])
+			} else {
+				// otherwise, add the item to the last group
+				last.push(item)
+			}
+			return previous
+		},
+		[[]]
+	)
+}
